+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Updated CLI dependencies and replaced `rand` with `getrandom` for generating random ids and ephemeral signing keys.
 
 ## [0.11.1] - 2026-02-08
 ### Changed